feat: filter plugins by surface
